BUILD: Update rpmbuild-docker el6 build image to use CentOS vault repos
CentOS 6 is EOL, standard YUM repos are not available anymore in mirrors.
This patch updates the docker base image used for el6 builds so it points
to the archive repos at vault.centos.org. As a consequence, the
prebuild.sh script gets removed since it was doing the same thing. The
difference is that the change in rpmbuild-docker will trigger the build
of a new el6 base image next time rpmbuild-docker -d el6 is run.