BUILD: check-updates: Allow json custom modification and filtering
When using json mode (second field of check-updates.def with a value
of "json"), raw data gets piped to jq <expression>, then piped to
grep -Pio (?<=")<pattern>(?=") with:
- default expression .
- default pattern [a-z+]+://.+

This was normally used to extract links to release tarballs.

With this patch, <expression> and <pattern> can be tunned in the
second field of check-updates.def as json:<expression>:<pattern>.

The patch also add support for "<" prefix in the fourth field. When
using it, the resulting pattern used to extract version numbers won't
start with a "/", but with a word boundary mark (\b or \< in regex).