BUILD: Use quictls release tags instead of branches
This is a change related to the detection of new releases of quictls
by the check-updates script. Branch OpenSSL_1_1_1w+quic is already
available but does not contain the quictls patches, so we cannot build
and release packages with OpenSSL_1_1_1w+quic yet. There is a pull
request on quictls github space but it has not been merged yet.
Let's use quictls release tags. Since there is no release tag for
OpenSSL_1_1_1w+quic so far, the check-update script won't alert
for a (non)available update of quictls.