BUG/MINOR: Do not use PROFILE=SYSTEM cipher list in quictls builds
Based on Github issue #4 (Does not support EL crypto-policies) reported
by nf-brentsaner, haproxy does not start with an ssl frontend using the
default haproxy.cfg provided in our quic package because it references
the PROFILE=SYSTEM cipher list.

That keyword requires a RedHat patch on openssl and we do not apply it
when building quictls. The patch replaces PROFILE=SYSTEM with HIGH in
default haproxy.cfg.