BUILD: Review default sysconfig file and keep it only in systemd builds
The init script shipped with this package on el < 7 is taken from
the source tarball in examples/haproxy.init. It does not source
/etc/sysconfig/haproxy so let's include it only in systemd builds.

The patch also modifies /etc/sysconfig/haproxy because its content was
not in coherent with the variables used in haproxy.service unit file.
Note that file is marked as %config(noreplace), so it won't be touched
during update.