BUILD: Use standard lua from the distribution
We used to build a recent version of lua here and static link haproxy
with it. It does not make sense to do that anymore since el6 and el7
support was dropped from the package. Both el8 and el9 have recent
enough versions of lua so let's get back to distro standard.